Removing duplicate code blocks by expanding the handler
