+++ src/MLP.py
@@ -5,6 +5,8 @@ reference: https://medium.com/biaslyai/pytorch-introduction-to-neural-network-fe
 import torch
 from src.Handler import *
 import src.CustomLoss as CustomLoss
+from torch.autograd import Variable
+
 
 
 class MLP(torch.nn.Module):
@@ -37,49 +39,3 @@ class MLPHandler(Handler):
 
     def create_model(self, input_shape, hidden_shape, output_shape):
         self.model = MLP(input_shape, hidden_shape, output_shape)
-
-    def train(self, x, y):
-        if self.loss_method == 'MSE':
-            criterion = CustomLoss.mse_loss
-        elif self.loss_method == 'Returns':
-            criterion = CustomLoss.return_loss
-        elif self.loss_method == 'Sharpe':
-            criterion = CustomLoss.sharpe_loss
-        else:
-            raise Exception("Invalid loss method")
-        optimizer = torch.optim.Adam(self.model.parameters(), lr=self.learning_rate)
-        self.model.eval()
-        self.model.train()
-        avg_losses = []
-
-        for epoch in range(self.epochs):
-            total_loss = 0
-            for i in range(0, x.shape[0], self.batch_size):
-                features = torch.FloatTensor(x[i:i+self.batch_size])
-                labels = torch.FloatTensor(y[i:i+self.batch_size])
-                l1reg = torch.tensor(0)
-                optimizer.zero_grad()
-                # Forward pass
-                y_pred = self.model(features)
-                # Compute Loss
-                loss = criterion(y_pred.squeeze(), labels)
-                if self.l1enable:
-                    for param in self.model.parameters():
-                        l1reg += torch.norm(param, 1).long()
-                    loss += self.alpha * l1reg
-
-                # Backward pass
-                loss.backward()
-                optimizer.step()
-                total_loss += loss.item()
-            avg_losses.append(total_loss / x.size)
-            print('epoch {}:\t loss {}'.format(epoch, total_loss / x.size))
-        return avg_losses
-
-    def test(self, x, y):
-        x = torch.FloatTensor(x)
-        y = torch.FloatTensor(y)
-        criterion = torch.nn.MSELoss()
-        y_pred = self.model(x)
-        loss = criterion(y_pred.squeeze(), y)
-        return loss, y_pred
